Fix java:S1905: Redundant casts should not be used

+++ src/main/java/charaplanner/ui/StatSlider.java
@@ -36,7 +36,7 @@ public class StatSlider extends VBox {
     }
 
     public void setValue(int value){
-        slider.setValue((double) value);
+        slider.setValue(value);
     }
 
     public int getValue(){
